NP-2157: Removed unneccesary annotations in model classes
Replaced them with class annotation
Moved Cristin model classes to own package
Created test for failure to enrich projects
Added charset to InputStreamReader
New json test data
