+++ src/main/java/no/unit/nva/cristin/projects/PresentationConverter.java
@@ -1,11 +1,19 @@
 package no.unit.nva.cristin.projects;
 
+//import static nva.commons.core.attempt.Try.attempt;
+
 import java.util.ArrayList;
 import java.util.Optional;
 import java.util.TreeMap;
 import java.util.stream.Collectors;
+import no.unit.nva.cristin.projects.model.cristin.CristinProject;
+//import nva.commons.core.JsonUtils;
+//import org.slf4j.Logger;
+//import org.slf4j.LoggerFactory;
 
 public class PresentationConverter {
+
+    //private static final Logger logger = LoggerFactory.getLogger(PresentationConverter.class);
 
     protected ProjectPresentation asProjectPresentation(CristinProject project, String language) {
         ProjectPresentation projectPresentation = new ProjectPresentation();
@@ -49,9 +57,15 @@ public class PresentationConverter {
                     fundingSourceNamePresentation.name = name.getValue();
                     return fundingSourceNamePresentation;
                 })
-                    .collect(Collectors.toList());
+                .collect(Collectors.toList());
             projectPresentation.fundings.add(fundingSourcePresentation);
         });
+
+        //String input = (attempt(() -> (JsonUtils.objectMapper.writeValueAsString(project))).get());
+        //logger.info(input);
+
+        //String output = (attempt(() -> (JsonUtils.objectMapper.writeValueAsString(projectPresentation))).get());
+        //logger.info(output);
 
         return projectPresentation;
     }
