+++ src/main/java/no/unit/nva/cristin/projects/model/cristin/FundingSource.java
@@ -1,0 +1,14 @@
+package no.unit.nva.cristin.projects.model.cristin;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import java.util.Map;
+
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public class FundingSource {
+
+    public String fundingSourceCode;
+    public String projectCode;
+    public Map<String, String> fundingSourceName;
+}
+
