+++ src/main/java/no/unit/nva/cristin/projects/model/cristin/Institution.java
@@ -1,0 +1,14 @@
+package no.unit.nva.cristin.projects.model.cristin;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import java.util.Map;
+
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public class Institution {
+
+    public String cristinInstitutionId;
+    public Map<String, String> institutionName;
+    public String url;
+}
+
