+++ src/main/java/no/unit/nva/cristin/projects/model/cristin/Role.java
@@ -1,16 +1,16 @@
-package no.unit.nva.cristin.projects;
+package no.unit.nva.cristin.projects.model.cristin;
 
-import static no.unit.nva.cristin.projects.JsonPropertyNames.ROLE_CODE;
 import static no.unit.nva.cristin.projects.JsonPropertyNames.UNIT;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @SuppressWarnings("PMD.ShortClassName")
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class Role {
 
-    @JsonProperty(ROLE_CODE)
     public String roleCode;
     public Institution institution;
-
     @JsonProperty(UNIT)
     public Unit institutionUnit;
