+++ src/main/java/no/unit/nva/cristin/projects/model/cristin/Unit.java
@@ -1,0 +1,16 @@
+package no.unit.nva.cristin.projects.model.cristin;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import java.util.Map;
+
+@SuppressWarnings("PMD.ShortClassName")
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public class Unit {
+
+    public String cristinUnitId;
+    public Map<String, String> unitName;
+    public String url;
+
+}
+
