+++ src/main/java/no/unit/nva/cristin/projects/model/nva/NvaProject.java
@@ -1,0 +1,7 @@
+package no.unit.nva.cristin.projects.model.nva;
+
+import nva.commons.core.JacocoGenerated;
+
+@JacocoGenerated
+public class NvaProject {
+}
