+++ src/test/resources/cristinQueryProjectsResponse.json
@@ -19,8 +19,8 @@
   {
     "cristin_project_id": "418126",
     "title": {
-      "no": "EALÁT - En studie av reindriftens sårbarhet og tilpasning under klimaendringer",
-      "en": "EALÁT - Reindeer Herders' Vulnerability Network Study: Reindeer Pastoralism in\r\na Changing Climate"
+      "no": "EALAT - En studie av reindriftens sårbarhet og tilpasning under klimaendringer",
+      "en": "EALAT - Reindeer Herders' Vulnerability Network Study: Reindeer Pastoralism in\r\na Changing Climate"
     },
     "main_language": "no",
     "url": "https://api.cristin.no/v2/projects/418126"
